fix animation glitches
