+++ app/sections/get-started/index.tsx
@@ -9,17 +9,16 @@ export const GetStarted = () => {
   return (
     <section className="flex flex-col items-center justify-center gap-10 my-20">
       <motion.div
-        initial={{ scale: 2 }}
+        initial={{ scale: 1.6 }}
         whileInView={{ scale: 1 }}
         transition={{ duration: 0.6 }}
-        className="relative bg-ramos-orange rounded-[35px] shadow-lg shadow-[#fe4a229a] h-[100px] w-[100px] flex justify-center items-center"
+        className="relative bg-ramos-orange rounded-[35px] shadow-lg shadow-[#fe4a229a] h-[95px] w-[95px] flex justify-center items-center"
       >
         <Image alt="chain" src="/images/chain.png" width={45} height={0} objectFit="cover" />
       </motion.div>
-
       <h2>
         <StaggeringAnimatedText
-          className="text-[#1A1A1A]  text-[90px] leading-[85px]"
+          className="text-ramos-black  text-[85px] leading-[70px]"
           text="Get Started"
         />
       </h2>
